Add files via upload
Title: "Building Modal Projects with JavaScript Classes.

This modal project work using / Manupulating  classes using javascript.
As a Passionate aspiring student developer, I've done this project.

Working:
When we click done on any of the show modal the we can able to display the content / modal .
In this I've used only single modal and  made then functionality for all 3 show modals using / manupulating the classes.
we can exit the displayed modal by clicking the Cros Mark (OR) By clicking on the out of the window (OR) We can exit by clicking on keyboard with Escape key.


Technology Used ; 
JavaScript:
I've  done modal project using java script for logic building with manipulating the classes. 

HTML:
The structure of web pages is built using HTML  to define content and layout.

CSS:
This is used to style the HTML elements, ensuring for cohesive design.

Conclution:

Hence,
This improve my skills in manupulating classes using java script. This design mostly used in e-commerce website.This became a great challenge for me.

👨🏻‍💻Happy coding...